refactor: 딜러와 플레이어 추상화
참가자들의 공통 로직 추상화
