+++ src/main/java/model/Participant.java
@@ -1,0 +1,16 @@
+package model;
+
+import java.util.List;
+
+public interface Participant {
+
+    int calculateTotalScore();
+
+    void addCard(Card card);
+
+    boolean canHit();
+
+    String getName();
+
+    List<Card> getCards();
+}
